requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
peewee==3.17.7
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
TgCrypto==1.2.5
urllib3==2.2.3

yadisk~=3.1.0